requirements.txt
Flask==2.1.0
google-cloud-speech==2.16.0
google-cloud-texttospeech==2.12.0
flask-cors==3.0.10
gunicorn==20.1.0
Werkzeug==2.0.1
google-cloud-language==2.3.1